MVVM 완성
