+++ src/hooks/useFetchExchangeInfo.ts
@@ -6,13 +6,12 @@ const useFetchExchangeInfo = (currency: Currency) => {
   const [isReady, setReady] = useState(false);
   const [exchangeInfo, setExchangeInfo] = useState<ExchangeInfo | null>(null);
 
-  const getEurInfo = async () => {
-    const info = await fetchExchangeInfo(currency);
-    setExchangeInfo(info);
-    setReady(true);
-  };
-
   useEffect(() => {
+    const getEurInfo = async () => {
+      const info = await fetchExchangeInfo(currency);
+      setExchangeInfo(info);
+      setReady(true);
+    };
     getEurInfo();
   }, []);
